- rule_confidence改为字典，加快查询速度 - 如果电影存在在规则里，就用规则预测，小于0.5视为不喜欢 - 输出进度提示
User 132859 is done. Progress is 96%. Used time is 264s. Remaining time is 11s.
User 134251 is done. Progress is 97%. Used time is 266s. Remaining time is 8s.
User 135648 is done. Progress is 98%. Used time is 269s. Remaining time is 5s.
User 137143 is done. Progress is 99%. Used time is 272s. Remaining time is 2s.
User 138493 is done. Progress is 100%. Used time is 275s. Remaining time is 0s.
accuracy is 0.4824741525631396.

#18